Refactor RuleAlteredJobCompletionDetectAlgorithm parameter
