+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-core/src/main/java/org/apache/shardingsphere/data/pipeline/core/job/FinishedCheckJob.java
@@ -87,6 +87,7 @@ public final class FinishedCheckJob implements SimpleJob {
                         sourceWritingStopAlgorithm.resumeSourceWriting(schemaName, jobId + "");
                     }
                 }
+                log.info("job {} finished", jobId);
                 // CHECKSTYLE:OFF
             } catch (final Exception ex) {
                 // CHECKSTYLE:ON
