+++ shardingsphere-test/shardingsphere-pipeline-test/src/test/java/org/apache/shardingsphere/data/pipeline/scenario/rulealtered/IdleRuleAlteredJobCompletionDetectAlgorithmTest.java
@@ -17,12 +17,12 @@
 
 package org.apache.shardingsphere.data.pipeline.scenario.rulealtered;
 
+import org.apache.shardingsphere.data.pipeline.api.detect.AllIncrementalTasksAlmostFinishedParameter;
 import org.apache.shardingsphere.data.pipeline.core.util.ReflectionUtil;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.mockito.junit.MockitoJUnitRunner;
 
@@ -34,6 +34,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public final class IdleRuleAlteredJobCompletionDetectAlgorithmTest {
@@ -41,7 +42,7 @@ public final class IdleRuleAlteredJobCompletionDetectAlgorithmTest {
     @Mock
     private Properties propsMock;
     
-    private IdleRuleAlteredJobCompletionDetectAlgorithm detectAlgorithm = new IdleRuleAlteredJobCompletionDetectAlgorithm();
+    private final IdleRuleAlteredJobCompletionDetectAlgorithm detectAlgorithm = new IdleRuleAlteredJobCompletionDetectAlgorithm();
     
     @Before
     public void setup() throws Exception {
@@ -51,28 +52,28 @@ public final class IdleRuleAlteredJobCompletionDetectAlgorithmTest {
     
     @Test(expected = IllegalArgumentException.class)
     public void assertInitFailNoIdleThresholdKey() {
-        Mockito.when(propsMock.containsKey(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn(false);
+        when(propsMock.containsKey(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn(false);
         detectAlgorithm.init();
     }
     
     @Test(expected = IllegalArgumentException.class)
     public void assertInitFailInvalidIdleThresholdKey() {
-        Mockito.when(propsMock.containsKey(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn(true);
-        Mockito.when(propsMock.getProperty(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn("@");
+        when(propsMock.containsKey(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn(true);
+        when(propsMock.getProperty(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn("@");
         detectAlgorithm.init();
     }
     
     @Test(expected = IllegalArgumentException.class)
     public void assertInitFailNegativeIdleThresholdKey() {
-        Mockito.when(propsMock.containsKey(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn(true);
-        Mockito.when(propsMock.getProperty(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn("-8");
+        when(propsMock.containsKey(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn(true);
+        when(propsMock.getProperty(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn("-8");
         detectAlgorithm.init();
     }
     
     @Test
     public void assertInitSuccess() {
-        Mockito.when(propsMock.containsKey(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn(true);
-        Mockito.when(propsMock.getProperty(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn("4");
+        when(propsMock.containsKey(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn(true);
+        when(propsMock.getProperty(IdleRuleAlteredJobCompletionDetectAlgorithm.IDLE_THRESHOLD_KEY)).thenReturn("4");
         detectAlgorithm.init();
     }
     
@@ -83,21 +84,25 @@ public final class IdleRuleAlteredJobCompletionDetectAlgorithmTest {
     
     @Test
     public void assertFalseOnNullIncrementalTasks() {
-        assertFalse(detectAlgorithm.allIncrementalTasksAlmostFinished(null));
+        AllIncrementalTasksAlmostFinishedParameter parameter = AllIncrementalTasksAlmostFinishedParameter.builder().build();
+        assertFalse(detectAlgorithm.allIncrementalTasksAlmostFinished(parameter));
     }
     
     @Test
     public void assertFalseOnEmptyIncrementalTasks() {
-        assertFalse(detectAlgorithm.allIncrementalTasksAlmostFinished(Collections.emptyList()));
+        AllIncrementalTasksAlmostFinishedParameter parameter = AllIncrementalTasksAlmostFinishedParameter.builder().incrementalTaskIdleMinutes(Collections.emptyList()).build();
+        assertFalse(detectAlgorithm.allIncrementalTasksAlmostFinished(parameter));
     }
     
     @Test
     public void assertFalseOnFewPendingIncrementalTasks() {
-        assertFalse(detectAlgorithm.allIncrementalTasksAlmostFinished(Arrays.asList(10L, 50L)));
+        AllIncrementalTasksAlmostFinishedParameter parameter = AllIncrementalTasksAlmostFinishedParameter.builder().incrementalTaskIdleMinutes(Arrays.asList(10L, 50L)).build();
+        assertFalse(detectAlgorithm.allIncrementalTasksAlmostFinished(parameter));
     }
     
     @Test
     public void assertTrueWhenAllIncrementalTasksAlmostFinished() {
-        assertTrue(detectAlgorithm.allIncrementalTasksAlmostFinished(Arrays.asList(60L, 50L, 30L)));
+        AllIncrementalTasksAlmostFinishedParameter parameter = AllIncrementalTasksAlmostFinishedParameter.builder().incrementalTaskIdleMinutes(Arrays.asList(60L, 50L, 30L)).build();
+        assertTrue(detectAlgorithm.allIncrementalTasksAlmostFinished(parameter));
     }
 }
